backend/requirements.txt
Django==3.2.12
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
